backend/requirements.txt
fastapi
uvicorn
firebase-admin
PyPDF2
python-dotenv
pydantic 
google-genai
markdown2
xhtml2pdf